+++ UUID.h
@@ -18,8 +18,10 @@
 #define UUID_LIB_VERSION              (F("0.2.0"))
 
 //  TODO an enum?
-const uint8_t UUID_MODE_VARIANT4 = 0;
-const uint8_t UUID_MODE_RANDOM   = 1;
+const uint8_t UUID_MODE_RANDOM   = 0;
+const uint8_t UUID_MODE_VARIANT4 = 4;  //  will become obsolete
+const uint8_t UUID_MODE_VERSION4 = 4;  //
+const uint8_t UUID_MODE_VERSION8 = 8;  //  not supported yet
 
 
 /////////////////////////////////////////////////
@@ -29,12 +31,15 @@ const uint8_t UUID_MODE_RANDOM   = 1;
 class UUID : public Printable
 {
 public:
+  //  seeds the UUID instance with compile time constants
+  //  plus the micros() value as run time component
+  //  to be fairly random for every instance.
   UUID();
 
   //  at least one seed value is mandatory, two is better.
   //  seed can be random or explicitly fixed.
   void     seed(uint32_t s1, uint32_t s2 = 0);
-  //  void     autoSeed();  //  TODO experimental 
+  //  void     autoSeed();  //  TODO experimental
 
   //  generate a new UUID
   void     generate();
@@ -42,12 +47,17 @@ public:
   char *   toCharArray();
 
   //  MODE
-  void     setVariant4Mode();
   void     setRandomMode();
+  void     setVersion4Mode();
+  //  void     setVersion8Mode(uint8_t nodeId[6]);
   int      getMode();
 
   //  Printable interface
   size_t   printTo(Print& p) const;
+
+  //  will be removed in 0.3.0
+  [[deprecated("Use setVersion4Mode() instead")]]
+  void     setVariant4Mode();
 
 
 private:
@@ -59,7 +69,8 @@ private:
 
   //  UUID in string format
   char     _buffer[37];
-  int _mode = UUID_MODE_VARIANT4;
+  //  default mode.
+  int      _mode = UUID_MODE_VARIANT4;
 };